2023/04/15

+++ ABC017_A.py
@@ -2,4 +2,4 @@ s1, e1 = map(int, input().split())
 s2, e2 = map(int, input().split())
 s3, e3 = map(int, input().split())
 
-print(s1*(e1/10)+s2*(e2/10)+s3*(e3/10))
+print(int(s1*(e1/10)+s2*(e2/10)+s3*(e3/10)))
